% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ReadCompDataSets.R
\name{ReadCompDataSets}
\alias{ReadCompDataSets}
\title{Read multiple expression data Sets that shall be compared}
\usage{
ReadCompDataSets()
}
\description{
This function
}
\author{
Hajk-Georg Drost
}

